aula-1/transactions/requirements.txt
beautifulsoup4==4.9.3
cassandra-driver==3.25.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
fastavro==1.4.1
geomet==0.2.1.post1
google==3.0.0
grpcio==1.38.0
grpcio-tools==1.38.0
idna==2.10
kafka-python==2.0.2
protobuf==3.17.3
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
